source/terraform/test_docker_files/tracking/requirements.txt
psycopg2-binary
dotenv